+++ lib/fog/dnsimple/requests/dns/delete_domain.rb
@@ -9,15 +9,10 @@ module Fog
         # DNSimple this will not delete the domain from the registry.
         #
         # ==== Parameters
-        # * account_id<~String> - the account the domain belong to
         # * zone_name<~String> - zone name
         #
         def delete_domain(zone_name)
-          request(
-            expects:  204,
-            method:   "DELETE",
-            path:     "/#{@dnsimple_account}/domains/#{zone_name}"
-          )
+          request { |client| client.domains.delete_domain(@dnsimple_account, zone_name) }
         end
       end