+++ test/test_helper.rb
@@ -8,7 +8,5 @@ require "fog/dnsimple"
 
 ENV["FOG_MOCK"] ||= "true"
 
-Excon.defaults.merge!(debug_request: true, debug_response: true)
-
 require_relative "helpers/fog_helper"
 require_relative "helpers/mock_helper"